`Refactored filters in product pages to use a single Filters component`
