+++ src/constants/index.ts
@@ -1,1 +1,2 @@
+export * from "./categories";
 export const DEFAULT_FILTER = { name: "All", value: "all", default: true };
